Adds API endpoint to retrieve patients associated with a client
A new API endpoint has been added to the `ClientesController` to retrieve all patients associated with a client based on their CIF (Client Identification Number). This endpoint allows clients to fetch their associated patients for further processing or display. The endpoint first searches for the client using the provided CIF, and if the client is found, it retrieves all the patients associated with that client. If no patients are found, a message indicating the absence of patients is returned. The response includes the client information and the list of associated patients. This change enhances the functionality of the system and improves the user experience for clients.

Fixes #789